use util instead of sentry service
