+++ packages/ssr-web/app/app.js
@@ -2,7 +2,7 @@ import Application from '@ember/application';
 import Resolver from 'ember-resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from '@cardstack/ssr-web/config/environment';
-import { initSentry } from './services/sentry';
+import { initSentry } from './utils/sentry';
 
 initSentry();
